cleaner version of "fat pivot"
Invariant is easier to reason about, since
there is always an "=" element in (p, i).
Also, we can avoid doing two swaps on each
item of the scan.

In addition, we return the endpoints of the
"=" interval, which we should have done
earlier, so we can avoid re-partitioning the
duplicate elements.